pyproject.toml
[project]
name = "S15A"
version = "0.2.0"
description = "Agentic AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.7.0",
    "pyyaml",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.3",
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "python-dotenv>=1.0.1",
    "faiss-cpu>=1.10.0",
    "httpx>=0.28.1",
    "llama-index>=0.12.31",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pymupdf4llm>=0.0.21",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "trafilatura[all]>=2.0.0",
    "jinja2>=3.1.6",
    "playwright>=1.52.0",
    "readability-lxml>=0.8.4.1",
    "astor>=0.8.1",
    "networkx>=3.4.2",
    "json-repair>=0.44.1",
    "spacy>=3.8.4",
    "markdownify>=1.1.0",
    "uvicorn>=0.34.2",
    "fastapi>=0.115.12",
    "sse-starlette>=2.3.4",
    "feedparser>=6.0.12",
    "beautifulsoup4>=4.13.4",
    "aiohttp>=3.11.18",
    "gradio>=6.7.0",
    "python-jose[cryptography]>=3.5.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "apscheduler>=3.10.0",
]
